Moved response logging to a common package
