--- dpc-attribution/src/main/java/gov/cms/dpc/attribution/ResponseLoggingFilter.java
@@ -1,31 +1,0 @@
-package gov.cms.dpc.attribution;
-
-import com.google.inject.Provider;
-import gov.cms.dpc.common.Constants;
-import gov.cms.dpc.common.MDCConstants;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.slf4j.MDC;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.ws.rs.container.ContainerRequestContext;
-import javax.ws.rs.container.ContainerResponseContext;
-import javax.ws.rs.container.ContainerResponseFilter;
-import javax.ws.rs.core.Context;
-
-public class ResponseLoggingFilter implements ContainerResponseFilter {
-
-    @Context
-    private Provider<HttpServletRequest> request;
-
-    private static final Logger logger = LoggerFactory.getLogger(ResponseLoggingFilter.class);
-
-    @Override
-    public void filter(ContainerRequestContext requestContext, ContainerResponseContext responseContext){
-        String requestId = MDC.get(MDCConstants.DPC_REQUEST_ID);
-        if(requestId!=null){
-            responseContext.getHeaders().add(Constants.DPC_REQUEST_ID_HEADER, requestId);
-        }
-        logger.info("resource_requested={}, method={}, status={}",requestContext.getUriInfo().getPath(),requestContext.getMethod(),responseContext.getStatus());
-    }
-}
